Cache: Add and implement StreamCache to replace BinaryCache
Why? Because loading data from file directly to memory and save to file from it - unefficient.
Also, this approach could lead to OutOfMemory situation (which made me to refactor cache system)

Streams on other hand "redirects" a data to file/whatever and doesn't store all of the data in memory, only very small amount of it (8kb buffer).
Furthermore, this approach is a little faster, but it's not much of a difference